requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.8
filelock==3.12.3
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.9.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
ifnude==0.0.3
imageio==2.31.5
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.14.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
nudenet==3.0.5
nudity==0.2.3
numpy==1.25.2
oauthlib==3.2.2
onnx==1.14.1
onnx-tf==1.10.0
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
parsel==1.8.1
Pillow==10.0.0
Protego==0.3.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt6==6.5.2
PyQt6-3D==6.5.0
PyQt6-Charts==6.5.0
PyQt6-DataVisualization==6.5.0
PyQt6-NetworkAuth==6.5.0
PyQt6-sip==13.5.2
PyQt6-WebEngine==6.5.0
pyqtgraph==0.13.3
python-dateutil==2.8.2
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.11.3
Scrapy==2.10.1
service-identity==23.1.0
setproctitle==1.3.2
six==1.16.0
soupsieve==2.5
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-addons==0.21.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-probability==0.22.0
termcolor==2.3.0
tifffile==2023.9.26
tldextract==3.5.0
tqdm==4.66.1
Twisted==22.10.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==2.0.4
w3lib==2.1.2
Werkzeug==3.0.0
wrapt==1.14.1
zope.interface==6.0
